# specs/001-saas-gate-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Gate Management API
  version: 1.0.0
  description: |
    Multi-tenant gate access management system API.
    All tenant endpoints require JWT authentication with tenant context.
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh token)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== SUPER ADMIN ====================
  /admin/plans:
    get:
      tags: [Super Admin]
      summary: List subscription plans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionPlan'
    post:
      tags: [Super Admin]
      summary: Create subscription plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlan'
...
  /admin/tenants:
    get:
      tags: [Super Admin]
      summary: List all tenants
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, trial, expired, suspended]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated tenant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTenants'
    post:
      tags: [Super Admin]
      summary: Onboard new tenant (building)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardTenantRequest'
      responses:
        '201':
          description: Tenant created with admin user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantWithAdmin'
  /admin/tenants/{tenantId}:
    get:
      tags: [Super Admin]
      summary: Get tenant details
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
    patch:
      tags: [Super Admin]
      summary: Update tenant
      security:
        - bearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant updated

  # ==================== USERS ====================
...
  /gates:
    get:
      tags: [Gates]
      summary: List gates in tenant
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Gate list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gate'
    post:
      tags: [Gates]
      summary: Create gate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGateRequest'
      responses:
        '201':
          description: Gate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
  /gates/{gateId}:
    get:
      tags: [Gates]
      summary: Get gate details
      security:
        - bearerAuth: []
      parameters:
        - name: gateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gate with controller info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateWithController'
    patch:
      tags: [Gates]
      summary: Update gate
      security:
        - bearerAuth: []
      parameters:
        - name: gateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGateRequest'
      responses:
        '200':
          description: Gate updated
    delete:
      tags: [Gates]
      summary: Delete gate
      security:
        - bearerAuth: []
      parameters:
        - name: gateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gate deleted
...
  /events:
    get:
      tags: [Events]
      summary: Query access events
      security:
        - bearerAuth: []
      parameters:
        - name: gateId
          in: query
          schema:
            type: string
            format: uuid
        - name: method
          in: query
          schema:
            type: string
            enum: [car_rfid, human_rfid, qr, web_app, manual]
        - name: result
          in: query
          schema:
            type: string
            enum: [allowed, denied]
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEvents'
  /events/export:
    get:
      tags: [Events]
      summary: Export events as CSV
      security:
        - bearerAuth: []
      parameters:
        - name: gateId
          in: query
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary

  # ==================== POLICIES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    # ===== USER =====
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
          enum: [super_admin, building_admin, security, resident_owner, resident_tenant, staff_maid, staff_driver]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        unitNumber:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        subscriptionPlanId:
          type: string
          format: uuid
        subscriptionStatus:
          type: string
          enum: [active, trial, expired, suspended]
        subscriptionExpiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TenantWithAdmin:
      allOf:
        - $ref: '#/components/schemas/Tenant'
        - type: object
          properties:
            adminUser:
              $ref: '#/components/schemas/User'
            temporaryPassword:
              type: string
    OnboardTenantRequest:
      type: object
      required: [name, slug, contactEmail, subscriptionPlanId, adminEmail, adminFirstName, adminLastName]
      properties:
        name:
          type: string
        slug:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        subscriptionPlanId:
          type: string
          format: uuid
        adminEmail:
          type: string
        adminFirstName:
          type: string
        adminLastName:
          type: string
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        subscriptionPlanId:
          type: string
          format: uuid
        subscriptionStatus:
          type: string
    PaginatedTenants:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tenant'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ===== SUBSCRIPTION PLAN =====
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        maxGates:
          type: integer
        maxUsers:
          type: integer
        logRetentionDays:
          type: integer
        features:
          type: object
        priceMonthly:
          type: number
        isActive:
          type: boolean
    CreatePlanRequest:
      type: object
      required: [name, maxGates, maxUsers, logRetentionDays]
      properties:
        name:
          type: string
        maxGates:
          type: integer
        maxUsers:
          type: integer
        logRetentionDays:
          type: integer
        features:
          type: object
        priceMonthly:
          type: number
...
    Gate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [vehicle, pedestrian]
        currentState:
          type: string
          enum: [CLOSED, OPENING, OPEN, CLOSING, OBSTACLE_HOLD, FAULT, MANUAL_OVERRIDE]
        location:
          type: string
        isSimulationMode:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    GateWithController:
      allOf:
        - $ref: '#/components/schemas/Gate'
        - type: object
          properties:
            controller:
              $ref: '#/components/schemas/GateController'
    GateController:
      type: object
      properties:
        id:
          type: string
          format: uuid
        deviceId:
          type: string
        firmwareVersion:
          type: string
        connectionStatus:
          type: string
          enum: [online, offline, unknown]
        lastHeartbeatAt:
          type: string
          format: date-time
        wifiStrength:
          type: integer
        uptimeSeconds:
          type: integer
    CreateGateRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [vehicle, pedestrian]
        location:
          type: string
    UpdateGateRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        isActive:
          type: boolean
...
    AccessEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gateId:
          type: string
          format: uuid
        gateName:
          type: string
        timestamp:
          type: string
          format: date-time
        method:
          type: string
          enum: [car_rfid, human_rfid, qr, web_app, manual]
        subjectType:
          type: string
          enum: [vehicle, rfid_card, visitor_pass, user, unknown]
        subjectId:
          type: string
          format: uuid
        subjectName:
          type: string
        result:
          type: string
          enum: [allowed, denied]
        denialReason:
          type: string
        operatorId:
          type: string
          format: uuid
        operatorName:
          type: string
        previousState:
          type: string
        newState:
          type: string
    PaginatedEvents:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AccessEvent'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ===== POLICIES =====
...